+++ packages/core/src/testing/mocks.ts
@@ -341,6 +341,7 @@ export const mockModule = (mod: Partial<d.Module> = {}): d.Module => ({
   cmps: [],
   isExtended: false,
   isMixin: false,
+  hasExportableMixins: false,
   coreRuntimeApis: [],
   outputTargetCoreRuntimeApis: {},
   collectionName: '',