+++ test/wdio/ts-target/extends-external-abstract/cmp.tsx
@@ -1,0 +1,41 @@
+import { Component, h, Prop, State, Method, Watch } from '@stencil/core';
+import { SiblingAbstractMixin } from 'test-sibling/dist/collection/sibling-abstract-mixin/sibling-abstract-mixin';
+
+/**
+ * A component that extends from an external library's abstract mixin class.
+ * This tests Bug B: importing abstract mixin classes from a lib - those classes'
+ * members should be properly merged in and have reactivity.
+ */
+@Component({
+  tag: 'extends-external-abstract',
+})
+export class ExtendsExternalAbstract extends SiblingAbstractMixin {
+  @Prop() prop1: string = 'default text';
+  @Watch('prop1')
+  prop1Changed(newValue: string) {
+    console.info('main class handler prop1:', newValue);
+  }
+
+  @State() state1: string = 'default state text';
+  @Watch('state1')
+  state1Changed(newValue: string) {
+    console.info('main class handler state1:', newValue);
+  }
+
+  @Method()
+  async method1() {
+    this.prop1 = 'main class method1 called';
+  }
+
+  render() {
+    return (
+      <div>
+        <p class="main-prop-1">Main class prop1: {this.prop1}</p>
+        <p class="main-prop-2">Main class prop2: {this.prop2}</p>
+        <p class="main-getter-prop">Main class getterProp: {this.getterProp}</p>
+        <p class="main-state-1">Main class state1: {this.state1}</p>
+        <p class="main-state-2">Main class state2: {this.state2}</p>
+      </div>
+    );
+  }
+}